successfully replaced seperate layout components since each is only being used by one page. Reducing confusion
