+++ sass/main.scss
@@ -27,10 +27,8 @@
 // Pages
 @import 'pages/app';
 @import 'pages/contact';
-
-// Layouts
-@import 'layouts/maximum';
-@import 'layouts/minimal';
+@import 'pages/blog';
+@import 'pages/sites';
 
 // Sites
 @import 'sites/luxotus-blog';
